requirements.txt
streamlit
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-core
openai
python-dotenv
langchain_google_genai 
pypdf
-e .